# docs/openapi.yml
openapi: 3.0.1
info:
  title: nfcwalker
  version: 1.0.0
paths:
...
  /api/incidents:
    get:
      operationId: listIncidents
      parameters:
      - name: organizationId
        in: query
        required: true
        explode: false
        schema:
          type: string
          format: uuid
      - name: siteId
        in: query
        explode: false
        schema:
          type: string
          format: uuid
          nullable: true
      - name: status
        in: query
        explode: false
        schema:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/IncidentStatus"
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: listIncidents 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_IncidentResponse_"
    post:
      operationId: createIncident
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - metadata
              type: object
              properties:
                metadata:
                  $ref: "#/components/schemas/IncidentCreateRequest"
                photos:
                  type: array
                  nullable: true
                  items:
                    type: string
                    format: binary
            encoding:
              metadata:
                contentType: application/octet-stream
                explode: true
              photos:
                contentType: application/octet-stream
                explode: true
        required: true
      responses:
        "200":
          description: createIncident 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentResponse"
  /api/incidents/{id}:
    get:
      operationId: getIncident
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getIncident 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentResponse"
    delete:
      operationId: deleteIncident
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: deleteIncident 200 response
    patch:
      operationId: patchIncident
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentPatchRequest"
        required: true
      responses:
        "200":
          description: patchIncident 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentResponse"
  /api/incidents/{id}/attach-scan/{scanEventId}:
    post:
      operationId: attachToScan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: scanEventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: attachToScan 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentResponse"
  /api/incidents/{id}/photos:
    post:
      operationId: addPhotosToIncident
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  nullable: true
                  items:
                    type: string
                    format: binary
            encoding:
              photos:
                contentType: application/octet-stream
                explode: false
        required: true
      responses:
        "200":
          description: addPhotosToIncident 200 response
          content:
            application/json:
              schema:
                type: object
  /api/incidents/{id}/photos/{photoId}:
    delete:
      operationId: deletePhotoFromIncident
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: deletePhotoFromIncident 200 response
  /api/invitations:
    get:
      operationId: getInvitations
      parameters:
      - name: organizationId
        in: query
        required: true
        explode: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: getInvitations 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_InvitationResponse_"
    post:
      operationId: createInvitation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
        required: true
      responses:
        "200":
          description: createInvitation 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
  /api/invitations/{id}:
    get:
      operationId: getInvitation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getInvitation 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
    delete:
      operationId: cancelInvitation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: cancelInvitation 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/organizations:
    get:
      operationId: listOrganizations
      parameters:
      - name: page
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: listOrganizations 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_OrganizationResponse_"
    post:
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
        required: true
      responses:
        "200":
          description: createOrganization 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
  /api/organizations/{id}:
    get:
      operationId: getOrganization
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: getOrganization 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
    put:
      operationId: updateOrganization
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
        required: true
      responses:
        "200":
          description: updateOrganization 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
    delete:
      operationId: deleteOrganization
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: deleteOrganization 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
...
components:
  schemas:
...
    CreateInvitationRequest:
      required:
      - email
      - organizationId
      - role
      type: object
      properties:
        email:
          type: string
        organizationId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/Role"
    CreateOrganizationRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
...
    IncidentCreateRequest:
      required:
      - description
      - severity
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
          nullable: true
        siteId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
        severity:
          $ref: "#/components/schemas/IncidentSeverity"
        checkpointId:
          type: string
          format: uuid
          nullable: true
        scanEventId:
          type: string
          format: uuid
          nullable: true
    IncidentPatchRequest:
      type: object
      properties:
        description:
          type: string
          nullable: true
        severity:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/IncidentSeverity"
        status:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/IncidentStatus"
    IncidentResponse:
      required:
      - createdAt
      - description
      - id
      - organizationId
      - reportedBy
      - severity
      - siteId
      - status
      - updatedAt
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        checkpointId:
          type: string
          format: uuid
          nullable: true
        scanEventId:
          type: string
          format: uuid
          nullable: true
        reportedBy:
          type: string
          format: uuid
        description:
          type: string
        severity:
          $ref: "#/components/schemas/IncidentSeverity"
        status:
          $ref: "#/components/schemas/IncidentStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IncidentSeverity:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
    IncidentStatus:
      type: string
      enum:
      - OPEN
      - IN_PROGRESS
      - RESOLVED
      - CLOSED
    InvitationResponse:
      required:
      - email
      - expiresAt
      - id
      - organizationId
      - role
      - status
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        organizationId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/InvitationStatus"
        expiresAt:
          type: string
          format: date-time
    InvitationStatus:
      type: string
      enum:
      - PENDING
      - ACCEPTED
      - EXPIRED
      - REVOKED
    OrganizationResponse:
      required:
      - createdAt
      - id
      - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Page_IncidentResponse_:
      required:
      - totalPages
      - totalSize
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_IncidentResponse_"
      - type: object
        properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
    Page_InvitationResponse_:
      required:
      - totalPages
      - totalSize
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_InvitationResponse_"
      - type: object
        properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
    Page_OrganizationResponse_:
      required:
      - totalPages
      - totalSize
      type: object
      allOf:
      - $ref: "#/components/schemas/Slice_OrganizationResponse_"
      - type: object
        properties:
          totalSize:
            type: integer
            format: int64
          totalPages:
            type: integer
            format: int32
...
    Pageable:
      required:
      - mode
      - number
      - size
      - sort
      type: object
      allOf:
      - $ref: "#/components/schemas/Sort"
      - type: object
        properties:
          number:
            type: integer
            format: int32
          size:
            type: integer
            format: int32
          mode:
            $ref: "#/components/schemas/Pageable.Mode"
          sort:
            $ref: "#/components/schemas/Sort"
    Pageable.Mode:
      type: string
      enum:
      - CURSOR_NEXT
      - CURSOR_PREVIOUS
      - OFFSET
...
    Role:
      type: string
      enum:
      - ROLE_APP_OWNER
      - ROLE_BOSS
      - ROLE_WORKER
...
    Slice_IncidentResponse_:
      required:
      - content
      - empty
      - numberOfElements
      - offset
      - pageNumber
      - pageable
      - size
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/IncidentResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
    Slice_InvitationResponse_:
      required:
      - content
      - empty
      - numberOfElements
      - offset
      - pageNumber
      - pageable
      - size
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/InvitationResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
    Slice_OrganizationResponse_:
      required:
      - content
      - empty
      - numberOfElements
      - offset
      - pageNumber
      - pageable
      - size
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        empty:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
...
    Sort:
      required:
      - orderBy
      type: object
      properties:
        orderBy:
          type: array
          items:
            $ref: "#/components/schemas/Sort.Order"
    Sort.Order:
      required:
      - ascending
      - direction
      - ignoreCase
      - property
      type: object
      properties:
        ignoreCase:
          type: boolean
        direction:
          $ref: "#/components/schemas/Sort.Order.Direction"
        property:
          type: string
        ascending:
          type: boolean
    Sort.Order.Direction:
      type: string
      enum:
      - ASC
      - DESC
...
    UpdateOrganizationRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
...